add user service api docu

+++ user_service/handler/user_handler.go
@@ -3,8 +3,11 @@ package handler
 import (
 	"encoding/json"
 	"net/http"
+	"os"
 	"strconv"
 
+	"user_services/errs"
+	"user_services/logs"
 	"user_services/security"
 	"user_services/service"
 
@@ -14,33 +17,44 @@ import (
 func (h userHandler) LoginUserHandler(w http.ResponseWriter, r *http.Request) {
 	var UserLoginRequese service.UserLoginRequest
 	if err := json.NewDecoder(r.Body).Decode(&UserLoginRequese); err != nil {
+		logs.Error(err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
 
 	user, err := h.userService.LoginUser(&UserLoginRequese)
 	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 
 	// Generate JWT token here if needed and add it to the response
-	token, err := security.NewBcryptHasher(h.secretKey).GenerateJWT(strconv.Itoa(user.ID))
+	token, err := security.NewBcryptHasher(os.Getenv("SECRET_KEY")).GenerateJWT(strconv.Itoa(user.ID))
 	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		logs.Error("Failed to generate JWT token: ")
+		handlerError(w, err)
 		return
 	}
 
 	w.Header().Set("Content-Type", "application/json")
 	w.Header().Set("Authorization", "Bearer "+token)
 	if err := json.NewEncoder(w).Encode(user); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		logs.Error("Failed to encode user response: ")
+		handlerError(w, err)
 		return
 	}
 }
 
 func (h userHandler) GetUserHandler(w http.ResponseWriter, r *http.Request) {
 	id, _ := strconv.Atoi(mux.Vars(r)["userId"])
+
+	// Authentication check
+	status := authenticationHandler(r)
+	if !status {
+		handlerError(w, errs.NewForbiddenError("forbidden"))
+		logs.Error("Authentication failed")
+		return
+	}
 
 	user, err := h.userService.GetUser(id)
 	if err != nil {
@@ -56,30 +70,40 @@ func (h userHandler) GetUserHandler(w http.ResponseWriter, r *http.Request) {
 }
 
 func (h userHandler) RegisterUserHandler(w http.ResponseWriter, r *http.Request) {
-	println("Have to register")
 	var UserRegisterRequese service.UserRegisterRequest
 	if err := json.NewDecoder(r.Body).Decode(&UserRegisterRequese); err != nil {
+		logs.Error(err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
 
 	status, err := h.userService.RegisterUser(UserRegisterRequese)
 	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 
 	w.Header().Set("Content-Type", "application/json")
 	if err := json.NewEncoder(w).Encode(status); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 }
 
 func (h userHandler) UpdateUserInfoHandler(w http.ResponseWriter, r *http.Request) {
 	id, _ := strconv.Atoi(mux.Vars(r)["userId"])
+
+	// Authentication check
+	status := authenticationHandler(r)
+	if !status {
+		handlerError(w, errs.NewForbiddenError("forbidden"))
+		logs.Error("Authentication failed")
+		return
+	}
+
 	var UserUpdateInfoRequese service.UserUpdateInfoRequest
 	if err := json.NewDecoder(r.Body).Decode(&UserUpdateInfoRequese); err != nil {
+		logs.Error(err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
@@ -87,21 +111,31 @@ func (h userHandler) UpdateUserInfoHandler(w http.ResponseWriter, r *http.Reques
 
 	user, err := h.userService.UpdateUserInfo(UserUpdateInfoRequese)
 	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 
 	w.Header().Set("Content-Type", "application/json")
 	if err := json.NewEncoder(w).Encode(user); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 }
 
 func (h userHandler) UpdateUserPasswordHandler(w http.ResponseWriter, r *http.Request) {
 	id, _ := strconv.Atoi(mux.Vars(r)["userId"])
+
+	// Authentication check
+	status := authenticationHandler(r)
+	if !status {
+		handlerError(w, errs.NewForbiddenError("forbidden"))
+		logs.Error("Authentication failed")
+		return
+	}
+
 	var UserUpdatePasswordRequese service.UserUpdatePasswordRequest
 	if err := json.NewDecoder(r.Body).Decode(&UserUpdatePasswordRequese); err != nil {
+		logs.Error(err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
@@ -109,21 +143,31 @@ func (h userHandler) UpdateUserPasswordHandler(w http.ResponseWriter, r *http.Re
 
 	user, err := h.userService.UpdateUserPassword(UserUpdatePasswordRequese)
 	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 
 	w.Header().Set("Content-Type", "application/json")
 	if err := json.NewEncoder(w).Encode(user); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 }
 
 func (h userHandler) UpdateUserEmailHandler(w http.ResponseWriter, r *http.Request) {
 	id, _ := strconv.Atoi(mux.Vars(r)["userId"])
+
+	// Authentication check
+	status := authenticationHandler(r)
+	if !status {
+		handlerError(w, errs.NewForbiddenError("forbidden"))
+		logs.Error("Authentication failed")
+		return
+	}
+
 	var UserUpdateEmailRequese service.UserUpdateEmailRequest
 	if err := json.NewDecoder(r.Body).Decode(&UserUpdateEmailRequese); err != nil {
+		logs.Error(err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
@@ -131,21 +175,31 @@ func (h userHandler) UpdateUserEmailHandler(w http.ResponseWriter, r *http.Reque
 
 	user, err := h.userService.UpdateUserEmail(UserUpdateEmailRequese)
 	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 
 	w.Header().Set("Content-Type", "application/json")
 	if err := json.NewEncoder(w).Encode(user); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 }
 
 func (h userHandler) UpdateUserUsernameHandler(w http.ResponseWriter, r *http.Request) {
 	id, _ := strconv.Atoi(mux.Vars(r)["userId"])
+
+	// Authentication check
+	status := authenticationHandler(r)
+	if !status {
+		handlerError(w, errs.NewForbiddenError("forbidden"))
+		logs.Error("Authentication failed")
+		return
+	}
+
 	var UserUpdateUsernameRequese service.UserUpdateUsernameRequest
 	if err := json.NewDecoder(r.Body).Decode(&UserUpdateUsernameRequese); err != nil {
+		logs.Error(err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
@@ -153,21 +207,31 @@ func (h userHandler) UpdateUserUsernameHandler(w http.ResponseWriter, r *http.Re
 
 	user, err := h.userService.UpdateUserUsername(UserUpdateUsernameRequese)
 	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 
 	w.Header().Set("Content-Type", "application/json")
 	if err := json.NewEncoder(w).Encode(user); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 }
 
 func (h userHandler) UpdateUserProfileHandler(w http.ResponseWriter, r *http.Request) {
 	id, _ := strconv.Atoi(mux.Vars(r)["userId"])
+
+	// Check authentication
+	status := authenticationHandler(r)
+	if !status {
+		handlerError(w, errs.NewForbiddenError("forbidden"))
+		logs.Error("Authentication failed")
+		return
+	}
+
 	var UserUpdateProfileRequese service.UserUpdateProfileRequest
 	if err := json.NewDecoder(r.Body).Decode(&UserUpdateProfileRequese); err != nil {
+		logs.Error(err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
@@ -175,13 +239,13 @@ func (h userHandler) UpdateUserProfileHandler(w http.ResponseWriter, r *http.Req
 
 	user, err := h.userService.UpdateUserProfile(UserUpdateProfileRequese)
 	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 
 	w.Header().Set("Content-Type", "application/json")
 	if err := json.NewEncoder(w).Encode(user); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 }
@@ -189,15 +253,23 @@ func (h userHandler) UpdateUserProfileHandler(w http.ResponseWriter, r *http.Req
 func (h userHandler) DeleteUserHandler(w http.ResponseWriter, r *http.Request) {
 	id, _ := strconv.Atoi(mux.Vars(r)["userId"])
 
+	// Authentication check
+	s := authenticationHandler(r)
+	if !s {
+		handlerError(w, errs.NewForbiddenError("forbidden"))
+		logs.Error("Authentication failed")
+		return
+	}
+
 	status, err := h.userService.DeleteUser(id)
 	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 
 	w.Header().Set("Content-Type", "application/json")
 	if err := json.NewEncoder(w).Encode(status); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
+		handlerError(w, err)
 		return
 	}
 }
